neo-strip: Switch to more portable fast pin writes.
Switch to using "digitalWriteFast" instead of PIN_MAP to manupulate
pins with tight timing requirements for the NEO Pixel devices. This
works well on the Particle Core, and compiles on the Photon, but
doesn't work properly.

This still allows the library to compile for all Particle devices, and
it should still work well for digital and dot light strips.